+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ## Version 2 Scene Essentials
 * 2.0.0: Pulled utilities
-* 2.1.0: Setup Home Screen and Level scenes and made test scene in tiled
-* 2.2.0: Got level properly loaded in and created and added Load assets scene.
+* 2.1.0: Setup Home Screen, Level scenes, made test scene in tiled
+* 2.2.0: Got level properly loaded in, created, added Load assets scene
 
 ## Version 3
+* 3.0.0: Created Player class, added Physics, Added walk animation